Remove DistillateType & Add value() function
